pyproject.toml
[tool.poetry]
name = "tinyticker"
version = "0.5.6"
description = "A tiny Raspberry Pi powered ePaper ticker."
authors = ["Loic Coyle <loic.coyle@hotmail.fr>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/loiccoyle/tinyticker"
keywords = ["raspberry-pi", "ticker", "stock", "crypto", "epaper", "finance"]

[tool.poetry.dependencies]
python = ">=3.9"
spidev = "^3.5"
Pillow = "^10.0.1"
cryptocompare = "^0.7.5"
pandas = "^1.3.5"
matplotlib = "^3.5.1"
mplfinance = "^0.12.7-alpha.17"
Flask = "^2.0.2"
yfinance = "^0.2.0"
qrcode = "^7.3.1"
packaging = "^23.0"
numpy = "^1.26"
watchdog = "^4.0.0"
waitress = "^3.0.0"
gpiozero = "^2.0.1"

[tool.poetry.scripts]
tinyticker = 'tinyticker.__main__:main'
tinyticker-web = 'tinyticker.web.__main__:main'

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
